Refactor and update enumerations and mappings for WGPU4K
This commit refactors multiple enumerations to improve alignment with updated specifications, including introducing new enum values and modifying existing ones. It also updates type aliases and mapping functions to support new data types like `ULong`. Additionally, redundant logger usage is replaced with simpler debugging via `println`.

+++ wgpu4k/src/commonMain/kotlin/Types.kt
@@ -4,11 +4,11 @@ typealias GPUBufferDynamicOffset = ULong
 
 typealias GPUBufferUsageFlags = Int
 
-typealias GPUColorWriteFlags = Int
+typealias GPUColorWriteFlags = ULong
 
 typealias GPUDepthBias = Int
 
-typealias GPUFlagsConstant = UInt
+typealias GPUFlagsConstant = ULong
 
 typealias GPUIndex32 = UInt
