+++ wgpu4k/src/jsMain/kotlin/Buffer.js.kt
@@ -13,7 +13,7 @@ actual class Buffer(internal val handler: GPUBuffer) : AutoCloseable {
     actual val size: GPUSize64
         get() = handler.size
     actual val usage: Set<BufferUsage>
-        get() = BufferUsage.entries.filter { it.value.toUInt() and handler.usage != 0u }.toSet()
+        get() = BufferUsage.entries.filter { it.value and handler.usage != 0uL }.toSet()
     actual val mapState: BufferMapState
         get() = BufferMapState.of(handler.mapState) ?: error("fail to get MapState")
